Use Guix: Pin dependencies
It was discovered that builds no longer were functional on my current version of
Guix (25b83bd). That is, builds would fail indicating 'avr-ld: cannot find crtatmega32u4.o:
No such file or directory'. avr-toolchain has been updated a few times, and an old version is
required to build this; as such, dependencies are now pinned with a channels.scm file, and an
.envrc has been provided for convenience.

A latter commit should merge qmk_firmware upstream, which likely will resolve the build issue
with the latest version of avr-gcc.